+++ packages/clients/diplan/src/store/updateState/index.ts
@@ -1,5 +1,12 @@
-import { FeatureCollection } from 'geojson'
+import {
+  FeatureCollection,
+  Feature as GeoJsonFeature,
+  MultiPolygon,
+  Polygon,
+} from 'geojson'
 import { PolarActionContext } from '@polar/lib-custom-types'
+import { cleanCoords } from '@turf/clean-coords'
+import { unkinkPolygon } from '@turf/unkink-polygon'
 import { DiplanGetters, DiplanState, GeometryType } from '../../types'
 import { mergeToMultiGeometries } from './mergeToMultiGeometries'
 import { validateGeoJson } from './validateGeoJson'
@@ -7,6 +14,33 @@ import { enrichWithMetaServices } from './enrichWithMetaServices'
 
 let abortController: AbortController | null = null
 export const drawFeatureCollectionSource = 'plugin/draw/featureCollection'
+
+const reviseFeatureCollection = (
+  sourceFeatureCollection: FeatureCollection<GeometryType>
+) => {
+  let revisedFeatureCollection = cloneFeatureCollection(sourceFeatureCollection)
+
+  revisedFeatureCollection = {
+    ...revisedFeatureCollection,
+    features: revisedFeatureCollection.features.reduce(
+      (accumulator, feature) => {
+        if (['Polygon', 'MultiPolygon'].includes(feature.geometry.type)) {
+          accumulator.push(
+            ...unkinkPolygon(
+              cleanCoords(feature) as GeoJsonFeature<Polygon | MultiPolygon>
+            ).features
+          )
+        } else {
+          accumulator.push(cleanCoords(feature))
+        }
+        return accumulator
+      },
+      [] as GeoJsonFeature<GeometryType>[]
+    ),
+  }
+
+  return revisedFeatureCollection
+}
 
 // FeatureCollection is compatible to stupid clone
 export const cloneFeatureCollection = (
@@ -31,19 +65,26 @@ export const updateState = async ({
   const thisController = (abortController = new AbortController())
 
   // clone to prevent accidentally messing with the draw tool's data
-  let revisedFeatureCollection = cloneFeatureCollection(
-    rootGetters[drawFeatureCollectionSource]
-  )
+  let revisedFeatureCollection
+  try {
+    revisedFeatureCollection = reviseFeatureCollection(
+      rootGetters[drawFeatureCollectionSource]
+    )
+  } catch {
+    // if this revision fails, errors are catastrophic and unfixable
+    commit('setSimpleGeometryValidity', false)
+    commit('setRevisionInProgress', false)
+    console.warn(
+      `@polar/client-diplan: Geometries entered were not valid. This may e.g. result from pulling points of a polygon in edit mode together until it's point-shaped.`
+    )
+    return
+  }
 
   // merge first; relevant for both follow-up steps
   if (getters.configuration.mergeToMultiGeometries) {
+    // TODO: turf provides "union" and "combine" methods, probably just use them
     revisedFeatureCollection = mergeToMultiGeometries(revisedFeatureCollection)
   }
-
-  // TODO consider from turf:
-  // * cleanCoords
-  // * union? or even better, combine?
-  // * unkinkPolygon?
 
   if (getters.configuration.validateGeoJson) {
     commit(