+++ packages/plugins/AddressSearch/CHANGELOG.md
@@ -1,8 +1,10 @@
 # CHANGELOG
 
-## unpublished
+## 3.1.0
 
 - Feature: `BKGParameters` has been exposed as root export.
+- Feature: Add new configuration parameter `component` to be able to interchange the UI component of this plugin.
+- Feature: Expose `SearchContainer.vue` as a new export in order to use it in custom implementations.
 
 ## 3.0.0